// src/app/Models/PostData_API_Class/SaveNewsClass.ts
import { showUploadFileClass } from "../showUploadFileClass";
import { uploadFileClass } from "../uploadFileClass";

// Generated by https://quicktype.io

export class SaveNewsClass {
  NewsID:        string;
  NewsHead:      string;
  NewsBody:      string;
  PostDate:      string;
  PostDeadline:  string;
  IsOn:          boolean;
  Sort:          number;
  KeyMan:        string;
  UploadFileOld: showUploadFileClass[];
  UploadFileNew: uploadFileClass[];
}